% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap_warnings.R
\name{aggregate_shap_values}
\alias{aggregate_shap_values}
\title{Aggregate SHAP Values from List}
\usage{
aggregate_shap_values(shap_results, feature_names)
}
\arguments{
\item{shap_results}{List of SHAP data.frames}

\item{feature_names}{Character vector of feature names}
}
\value{
Matrix of SHAP values (samples x features)
}
\description{
Aggregates SHAP results into a matrix.
}
\keyword{internal}
