% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_stats}
\alias{cache_stats}
\title{Get Cache Statistics}
\usage{
cache_stats(cache)
}
\arguments{
\item{cache}{Cache object}
}
\value{
A list with cache statistics
}
\description{
Returns information about cache usage.
}
