% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-differential-expression.R
\name{calculate_effect_sizes}
\alias{calculate_effect_sizes}
\title{Calculate Effect Sizes}
\usage{
calculate_effect_sizes(
  expression_data,
  case_idx,
  control_idx,
  effect_size_methods,
  verbose
)
}
\description{
Calculate Effect Sizes
}
\keyword{internal}
