% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-helpers.R
\name{calculate_selection_metrics}
\alias{calculate_selection_metrics}
\title{Calculate Feature Selection Metrics}
\usage{
calculate_selection_metrics(selected_features, de_results)
}
\arguments{
\item{selected_features}{List of feature sets from different methods}

\item{de_results}{Data frame of differential expression results}
}
\value{
Data frame with selection metrics
}
\description{
Calculate Feature Selection Metrics
}
\keyword{internal}
