% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepLearners.R
\name{check_torch_packages}
\alias{check_torch_packages}
\title{Check Deep Learning Dependencies}
\usage{
check_torch_packages()
}
\value{
TRUE if available, stops with error message otherwise
}
\description{
Checks if torch and mlr3torch are available.
}
\keyword{internal}
