% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics-select-main.R, R/omicselector-optimized.R
\name{cleanup_parallel_processing}
\alias{cleanup_parallel_processing}
\title{Cleanup Parallel Processing}
\usage{
cleanup_parallel_processing(parallel_setup)

cleanup_parallel_processing(parallel_setup)
}
\arguments{
\item{parallel_setup}{List from setup_parallel_processing}
}
\description{
Cleanup Parallel Processing

Cleanup Parallel Processing
}
\keyword{internal}
