% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterGOF.R
\name{compare_gof_filters}
\alias{compare_gof_filters}
\title{Compare GOF Filters on Task}
\usage{
compare_gof_filters(task, top_n = 20)
}
\arguments{
\item{task}{An mlr3 classification task}

\item{top_n}{Number of top features to compare}
}
\value{
A data frame with filter rankings
}
\description{
Compares different GOF filter rankings on a task.
}
