% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{compute_stability_from_resample}
\alias{compute_stability_from_resample}
\title{Compute Stability from ResampleResult}
\usage{
compute_stability_from_resample(
  resample_result,
  all_features,
  filter_id = "filter"
)
}
\arguments{
\item{resample_result}{An mlr3 ResampleResult object}

\item{all_features}{Character vector of all candidate features}

\item{filter_id}{The ID of the filter PipeOp (default: "filter")}
}
\value{
A NogueiraStability object
}
\description{
Convenience function to compute Nogueira Stability directly from
an mlr3 ResampleResult.
}
