% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_OmicSelector_improved.R
\name{convert_to_legacy_formulas}
\alias{convert_to_legacy_formulas}
\title{Convert Selected Features to Legacy Formula Format}
\usage{
convert_to_legacy_formulas(selected_features, class_column = "Class")
}
\arguments{
\item{selected_features}{List of feature vectors by method}

\item{class_column}{Character name of class column}
}
\value{
List of formulas compatible with legacy code
}
\description{
Convert Selected Features to Legacy Formula Format
}
\keyword{internal}
