% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-mirna-correction.R
\name{correct_names_parallel}
\alias{correct_names_parallel}
\title{Correct Names in Parallel}
\usage{
correct_names_parallel(
  column_names,
  mirbase_data,
  species,
  correct_dots,
  verbose
)
}
\description{
Correct Names in Parallel
}
\keyword{internal}
