% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-mirna-correction.R
\name{correct_names_sequential}
\alias{correct_names_sequential}
\title{Correct Names Sequentially}
\usage{
correct_names_sequential(
  column_names,
  mirbase_data,
  species,
  correct_dots,
  verbose
)
}
\description{
Correct Names Sequentially
}
\keyword{internal}
