% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-mirna-correction.R
\name{correct_single_name}
\alias{correct_single_name}
\title{Correct Single miRNA Name}
\usage{
correct_single_name(name, mirbase_data, species, correct_dots)
}
\description{
Correct Single miRNA Name
}
\keyword{internal}
