% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianTuner.R
\name{create_autotuner}
\alias{create_autotuner}
\title{Create AutoTuner with MBO}
\usage{
create_autotuner(learner, search_space, n_evals, inner_folds, measure)
}
\arguments{
\item{learner}{An mlr3 Learner}

\item{search_space}{A paradox ParamSet}

\item{n_evals}{Number of evaluations}

\item{inner_folds}{Inner CV folds}

\item{measure}{Measure name or object}
}
\value{
An AutoTuner object
}
\description{
Internal helper to create an AutoTuner with mlr3mbo.
}
\keyword{internal}
