% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-helpers.R
\name{create_feature_formula}
\alias{create_feature_formula}
\title{Create Feature Formula for Modeling}
\usage{
create_feature_formula(features, class_column = "Class")
}
\arguments{
\item{features}{Character vector of feature names}

\item{class_column}{Character name of class column (default: "Class")}
}
\value{
Formula object
}
\description{
Create Feature Formula for Modeling
}
\keyword{internal}
