% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap_warnings.R
\name{create_predict_function}
\alias{create_predict_function}
\title{Create Predict Function for Various Model Types}
\usage{
create_predict_function(model)
}
\arguments{
\item{model}{A model object}
}
\value{
A predict function
}
\description{
Creates a standardized predict function for different model types.
}
\keyword{internal}
