% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics-select-main.R
\name{create_summary_report}
\alias{create_summary_report}
\title{Create Summary Report}
\usage{
create_summary_report(results, wd)
}
\arguments{
\item{results}{OmicSelector results object}

\item{wd}{Working directory}
}
\description{
Create Summary Report
}
\keyword{internal}
