% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-utility-functions.R
\name{create_upset_plot}
\alias{create_upset_plot}
\title{Create UpSet Plot for Multi-way Intersections}
\usage{
create_upset_plot(signature_list)
}
\description{
Create UpSet Plot for Multi-way Intersections
}
\keyword{internal}
