% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplearning.R
\docType{package}
\name{deeplearning}
\alias{OmicSelector}
\alias{OmicSelector-package}
\alias{deeplearning}
\title{Deep Learning Functions for OmicSelector}
\description{
This module provides comprehensive deep learning capabilities for miRNA biomarker
analysis using Keras/TensorFlow neural networks with hyperparameter optimization,
autoencoder support, and transfer learning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biostat.umed.pl/OmicSelector/}
  \item \url{https://github.com/kstawiski/OmicSelector}
  \item Report bugs at \url{https://github.com/kstawiski/OmicSelector/issues}
}

}
\author{
\strong{Maintainer}: Konrad Stawiski \email{konrad.stawiski@umed.lodz.pl} (\href{https://orcid.org/0000-0002-6550-3384}{ORCID})

Authors:
\itemize{
  \item Marcin Kaszkowiak
}

Other contributors:
\itemize{
  \item Damian Mikulski [contributor]
}

}
