% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\title{Deprecated and Leaky Functions}
\description{
This file contains deprecation notices and warnings for functions
that violate the zero-leakage principle of OmicSelector 2.0.
}
