% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frozen-combat.R
\name{.compute_batch_effect}
\alias{.compute_batch_effect}
\title{Compute batch effect strength}
\usage{
.compute_batch_effect(data, batch)
}
\description{
Compute batch effect strength
}
\keyword{internal}
