% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch-learners.R
\name{.create_custom_torch_learner}
\alias{.create_custom_torch_learner}
\title{Create Custom Torch Learner}
\usage{
.create_custom_torch_learner(
  hidden_layers,
  dropout,
  activation,
  epochs,
  batch_size,
  lr,
  device
)
}
\description{
Create Custom Torch Learner
}
\keyword{internal}
