% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-export.R
\name{.create_export_readme}
\alias{.create_export_readme}
\title{Create export README}
\usage{
.create_export_readme(model_name, results, task, readme_path)
}
\description{
Create export README
}
\keyword{internal}
