% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch-learners.R
\name{.create_mlr3torch_mlp}
\alias{.create_mlr3torch_mlp}
\title{Create mlr3torch MLP (internal)}
\usage{
.create_mlr3torch_mlp(
  hidden_layers,
  dropout,
  activation,
  epochs,
  batch_size,
  lr,
  early_stopping,
  patience,
  device
)
}
\description{
Create mlr3torch MLP (internal)
}
\keyword{internal}
