% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{.deprecation_warning}
\alias{.deprecation_warning}
\title{Emit deprecation warning for leaky functions}
\usage{
.deprecation_warning(func_name, replacement = NULL, reason = NULL)
}
\arguments{
\item{func_name}{Function name}

\item{replacement}{Suggested replacement function}

\item{reason}{Reason for deprecation}
}
\description{
Issues a prominent warning when deprecated/leaky functions are called
}
\keyword{internal}
