% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-export.R
\name{.export_glmnet}
\alias{.export_glmnet}
\title{Export glmnet model (coefficients)}
\usage{
.export_glmnet(model, base_path, task, results)
}
\description{
Export glmnet model (coefficients)
}
\keyword{internal}
