% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature-selection.R
\name{.extract_consensus_from_bmr}
\alias{.extract_consensus_from_bmr}
\title{Extract Consensus Features from Benchmark Result}
\usage{
.extract_consensus_from_bmr(bmr, learner_id, min_frequency)
}
\description{
Extract Consensus Features from Benchmark Result
}
\keyword{internal}
