% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature-selection.R
\name{.extract_consensus_from_sf}
\alias{.extract_consensus_from_sf}
\title{Extract Consensus Features from Selected Features Per Fold}
\usage{
.extract_consensus_from_sf(sf, learner_id, min_frequency)
}
\description{
Extract Consensus Features from Selected Features Per Fold
}
\keyword{internal}
