% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-export.R
\name{.extract_estimator}
\alias{.extract_estimator}
\title{Extract estimator from GraphLearner or Learner}
\usage{
.extract_estimator(learner)
}
\description{
Extract estimator from GraphLearner or Learner
}
\keyword{internal}
