% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-export.R
\name{.extract_preprocessing_manifest}
\alias{.extract_preprocessing_manifest}
\title{Extract preprocessing manifest from GraphLearner}
\usage{
.extract_preprocessing_manifest(learner, task)
}
\description{
Extract preprocessing manifest from GraphLearner
}
\keyword{internal}
