% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-export.R
\name{.generate_plumber_api}
\alias{.generate_plumber_api}
\title{Generate Plumber API script}
\usage{
.generate_plumber_api(model_name, task, api_path)
}
\description{
Generate Plumber API script
}
\keyword{internal}
