% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.generate_simple_report}
\alias{.generate_simple_report}
\title{Generate Simple Text Report}
\usage{
.generate_simple_report(report_data, output_file)
}
\description{
Generate Simple Text Report
}
\keyword{internal}
