% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.populate_from_benchmark}
\alias{.populate_from_benchmark}
\title{Populate Report Data from Benchmark Result}
\usage{
.populate_from_benchmark(report_data, bmr)
}
\description{
Populate Report Data from Benchmark Result
}
\keyword{internal}
