% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{.populate_from_pipeline}
\alias{.populate_from_pipeline}
\title{Populate Report Data from Pipeline}
\usage{
.populate_from_pipeline(report_data, pipeline)
}
\description{
Populate Report Data from Pipeline
}
\keyword{internal}
