% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-omics.R
\name{.prepare_modality}
\alias{.prepare_modality}
\title{Prepare Single Modality Data}
\usage{
.prepare_modality(data, target, modality_name)
}
\description{
Prepare Single Modality Data
}
\keyword{internal}
