% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature-selection.R
\name{.select_weighted}
\alias{.select_weighted}
\title{Weighted Scoring Selection}
\usage{
.select_weighted(candidates, weights, na_stability)
}
\description{
Weighted Scoring Selection
}
\keyword{internal}
