% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature-selection.R
\name{.try_extract_feature_counts}
\alias{.try_extract_feature_counts}
\title{Try to Extract Feature Counts from Benchmark Result}
\usage{
.try_extract_feature_counts(candidates, bmr)
}
\description{
Try to Extract Feature Counts from Benchmark Result
}
\keyword{internal}
