% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-encodings.R
\name{encode_deepinsight}
\alias{encode_deepinsight}
\title{Encode with DeepInsight Layout}
\usage{
encode_deepinsight(x, layout)
}
\arguments{
\item{x}{Numeric feature vector (length p)}

\item{layout}{Output of \code{fit_deepinsight}}
}
\value{
A grid_size × grid_size numeric matrix
}
\description{
Encode with DeepInsight Layout
}
\examples{
layout <- list(positions = matrix(c(1, 1, 1, 2, 2, 1), ncol = 2, byrow = TRUE), grid_size = 2L)
encode_deepinsight(1:3, layout)

}
\references{
Sharma A et al. (2019). DeepInsight: A methodology to transform a
non-image data to an image for convolution neural network architecture.
\emph{Scientific Reports}, 9, 11399.
}
