% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-optimized.R
\name{execute_boruta_method}
\alias{execute_boruta_method}
\title{Execute Boruta Method}
\usage{
execute_boruta_method(train_data, class_column, max_iterations, verbose)
}
\description{
Execute Boruta Method
}
\keyword{internal}
