% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-optimized.R
\name{execute_cfs_method}
\alias{execute_cfs_method}
\title{Execute CFS Method}
\usage{
execute_cfs_method(train_data, class_column, verbose)
}
\description{
Execute CFS Method
}
\keyword{internal}
