% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-optimized.R
\name{execute_rf_rfe_method}
\alias{execute_rf_rfe_method}
\title{Execute Random Forest RFE Method}
\usage{
execute_rf_rfe_method(train_data, class_column, max_features, verbose)
}
\description{
Execute Random Forest RFE Method
}
\keyword{internal}
