% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-optimized.R
\name{execute_sa_method}
\alias{execute_sa_method}
\title{Execute Simulated Annealing Method}
\usage{
execute_sa_method(
  train_data,
  class_column,
  max_features,
  max_iterations,
  verbose
)
}
\description{
Execute Simulated Annealing Method
}
\keyword{internal}
