% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{extract_features_from_resample}
\alias{extract_features_from_resample}
\title{Extract Features from All Folds in ResampleResult}
\usage{
extract_features_from_resample(resample_result, filter_id = "filter")
}
\arguments{
\item{resample_result}{An mlr3 ResampleResult object}

\item{filter_id}{The ID of the filter PipeOp (default: "filter")}
}
\value{
A list of character vectors, one per fold
}
\description{
Extracts selected features from each fold of a ResampleResult.
}
