% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-helpers.R
\name{fast_cfs}
\alias{fast_cfs}
\title{Fast Implementation of Correlation-based Feature Selection}
\usage{
fast_cfs(data, class_column, max_features = 20)
}
\arguments{
\item{data}{Data frame with features and class}

\item{class_column}{Character name of class column}

\item{max_features}{Integer maximum features to return}
}
\value{
Character vector of selected features
}
\description{
Fast Implementation of Correlation-based Feature Selection
}
\keyword{internal}
