% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-helpers.R
\name{fast_ttest}
\alias{fast_ttest}
\title{Optimized t-test for Multiple Features}
\usage{
fast_ttest(feature_matrix, class_labels, case_label, control_label)
}
\arguments{
\item{feature_matrix}{Matrix of features (samples x features)}

\item{class_labels}{Factor of class labels}

\item{case_label}{Character positive case label}

\item{control_label}{Character negative control label}
}
\value{
Data frame with statistical results
}
\description{
Optimized t-test for Multiple Features
}
\keyword{internal}
