% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{features}
\alias{features}
\title{Extract Features from OmicSelector Object}
\usage{
features(x, method_id = NULL, ...)
}
\arguments{
\item{x}{OmicSelector object}

\item{method_id}{Optional specific method ID to extract features from}

\item{...}{Additional arguments (not used)}
}
\value{
List of character vectors containing selected features
}
\description{
Extract Features from OmicSelector Object
}
