% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda-feature-selection.R
\name{FilterCoDA_CoDaCoRe}
\alias{FilterCoDA_CoDaCoRe}
\title{CoDaCoRe Feature Filter}
\description{
`mlr3filters::Filter` wrapper around [codaFS_codacore_wrapper()].
}
\section{Dictionary}{

```
flt("coda_codacore")
mlr_filters$get("coda_codacore")
```
}

\section{Super class}{
\code{\link[mlr3filters:Filter]{mlr3filters::Filter}} -> \code{FilterCoDA_CoDaCoRe}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilterCoDA_CoDaCoRe-new}{\code{FilterCoDA_CoDaCoRe$new()}}
\item \href{#method-FilterCoDA_CoDaCoRe-clone}{\code{FilterCoDA_CoDaCoRe$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="calculate"><a href='../../mlr3filters/html/Filter.html#method-Filter-calculate'><code>mlr3filters::Filter$calculate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="format"><a href='../../mlr3filters/html/Filter.html#method-Filter-format'><code>mlr3filters::Filter$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="help"><a href='../../mlr3filters/html/Filter.html#method-Filter-help'><code>mlr3filters::Filter$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="print"><a href='../../mlr3filters/html/Filter.html#method-Filter-print'><code>mlr3filters::Filter$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterCoDA_CoDaCoRe-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilterCoDA_CoDaCoRe-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this Filter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterCoDA_CoDaCoRe$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterCoDA_CoDaCoRe-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilterCoDA_CoDaCoRe-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterCoDA_CoDaCoRe$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
