% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterGOF.R
\name{FilterGOF_KS}
\alias{FilterGOF_KS}
\title{Kolmogorov-Smirnov GOF Filter}
\description{
Uses the KS test D-statistic as feature importance score.
Captures differences in distribution shape, location, and zero-frequency
between two classes without assuming a parametric distribution.
}
\details{
The KS test compares Empirical Cumulative Distribution Functions (ECDF).
If one class has 80% zeros and the other has 20% zeros, the ECDFs show
a massive difference captured by the D statistic (bounded [0, 1]).

Higher D values indicate better class separation.
}
\section{Dictionary}{

This Filter can be instantiated via the dictionary `mlr_filters` or with
the associated sugar function `flt()`:
```
flt("gof_ks")
mlr_filters$get("gof_ks")
```
}

\references{
Kolmogorov, A. N. (1933). Sulla determinazione empirica di una legge di distribuzione.
}
\section{Super class}{
\code{\link[mlr3filters:Filter]{mlr3filters::Filter}} -> \code{FilterGOF_KS}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilterGOF_KS-new}{\code{FilterGOF_KS$new()}}
\item \href{#method-FilterGOF_KS-clone}{\code{FilterGOF_KS$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="calculate"><a href='../../mlr3filters/html/Filter.html#method-Filter-calculate'><code>mlr3filters::Filter$calculate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="format"><a href='../../mlr3filters/html/Filter.html#method-Filter-format'><code>mlr3filters::Filter$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="help"><a href='../../mlr3filters/html/Filter.html#method-Filter-help'><code>mlr3filters::Filter$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="print"><a href='../../mlr3filters/html/Filter.html#method-Filter-print'><code>mlr3filters::Filter$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterGOF_KS-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilterGOF_KS-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this Filter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterGOF_KS$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterGOF_KS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilterGOF_KS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterGOF_KS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
