% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterGOF.R
\name{FilterHurdle}
\alias{FilterHurdle}
\title{Hurdle Filter for Zero-Inflated Data}
\description{
Two-part filter designed for zero-inflated count/expression data.
Tests both the difference in zero-frequency AND the difference in
magnitude among non-zero values.
}
\details{
The Hurdle filter combines two tests:
1. **Binary component**: Chi-square test on zero vs non-zero contingency table
2. **Continuous component**: Wilcoxon test on non-zero values

The score is computed using Fisher's method to combine p-values:
\code{chi2 = -2 * sum(log(p_values))}

This captures genes that are good biomarkers because:
- They are detectable in one class but not the other (zero-frequency difference)
- They are upregulated in one class among expressed samples (magnitude difference)
}
\section{Dictionary}{

This Filter can be instantiated via the dictionary `mlr_filters` or with
the associated sugar function `flt()`:
```
flt("hurdle")
mlr_filters$get("hurdle")
```
}

\section{Super class}{
\code{\link[mlr3filters:Filter]{mlr3filters::Filter}} -> \code{FilterHurdle}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilterHurdle-new}{\code{FilterHurdle$new()}}
\item \href{#method-FilterHurdle-clone}{\code{FilterHurdle$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="calculate"><a href='../../mlr3filters/html/Filter.html#method-Filter-calculate'><code>mlr3filters::Filter$calculate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="format"><a href='../../mlr3filters/html/Filter.html#method-Filter-format'><code>mlr3filters::Filter$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="help"><a href='../../mlr3filters/html/Filter.html#method-Filter-help'><code>mlr3filters::Filter$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="print"><a href='../../mlr3filters/html/Filter.html#method-Filter-print'><code>mlr3filters::Filter$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterHurdle-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilterHurdle-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this Filter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterHurdle$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterHurdle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilterHurdle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterHurdle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
