% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-benchmarking.R
\name{find_best_models}
\alias{find_best_models}
\title{Find Best Models}
\usage{
find_best_models(results_matrix)
}
\arguments{
\item{results_matrix}{Results matrix}
}
\value{
List of best models by different metrics
}
\description{
Find Best Models
}
\keyword{internal}
