% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{formulas}
\alias{formulas}
\title{Extract Formulas from OmicSelector Object}
\usage{
formulas(x, ...)
}
\arguments{
\item{x}{OmicSelector object}

\item{...}{Additional arguments (not used)}
}
\value{
Character vector of formulas
}
\description{
Extract Formulas from OmicSelector Object
}
