% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-benchmarking.R
\name{get_data_info}
\alias{get_data_info}
\title{Get Data Information}
\usage{
get_data_info(data_list)
}
\arguments{
\item{data_list}{List with data frames}
}
\value{
Data information list
}
\description{
Get Data Information
}
\keyword{internal}
