% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-management.R
\name{get_default_config}
\alias{get_default_config}
\title{Get Default Configuration}
\usage{
get_default_config()
}
\value{
List of default configuration parameters
}
\description{
Returns hardcoded default configuration when config package is not available
}
\keyword{internal}
