% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-optimized.R
\name{get_feature_columns}
\alias{get_feature_columns}
\title{Get Feature Column Names}
\usage{
get_feature_columns(data, feature_prefix)
}
\arguments{
\item{data}{Data frame}

\item{feature_prefix}{Character prefix for feature columns}
}
\value{
Character vector of feature column names
}
\description{
Get Feature Column Names
}
\keyword{internal}
