% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-selection-pipeline.R,
%   R/omicselector-optimized.R
\name{get_method_name}
\alias{get_method_name}
\title{Get Method Name}
\usage{
get_method_name(method_id)

get_method_name(method_id)
}
\arguments{
\item{method_id}{Integer method identifier}

\item{method_num}{Integer method number}
}
\value{
Character string with method name

Character method name
}
\description{
Get Method Name

Get Method Name by ID
}
\keyword{internal}
