% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-omics.R
\name{get_modality_info}
\alias{get_modality_info}
\title{Get Modality Information}
\usage{
get_modality_info(omics_input)
}
\arguments{
\item{omics_input}{A validated OmicsInput object}
}
\value{
A data.frame summarizing each modality
}
\description{
Returns summary information about each modality in a multi-omics dataset.
}
