% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianTuner.R
\name{get_optimal_params}
\alias{get_optimal_params}
\title{Get Optimal Hyperparameters from AutoTuner}
\usage{
get_optimal_params(autotuner)
}
\arguments{
\item{autotuner}{A trained AutoTuner object}
}
\value{
A list of optimal hyperparameter values
}
\description{
Extracts the best hyperparameters found by Bayesian optimization.
}
