% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-helpers.R
\name{\%||\%}
\alias{\%||\%}
\title{Null Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left operand}

\item{y}{Right operand (default if x is NULL)}
}
\value{
x if not NULL, otherwise y
}
\description{
Null Coalescing Operator
}
\keyword{internal}
