% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoXAI.R
\name{identify_correlation_clusters}
\alias{identify_correlation_clusters}
\title{Identify Correlation Clusters}
\usage{
identify_correlation_clusters(high_cor_pairs, all_features)
}
\arguments{
\item{high_cor_pairs}{Data frame from xai_correlations}

\item{all_features}{All feature names}
}
\value{
List of character vectors (clusters)
}
\description{
Groups highly correlated features into clusters.
}
\keyword{internal}
