% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-optimized.R
\name{initialize_omicselector_results}
\alias{initialize_omicselector_results}
\title{Initialize OmicSelector Results Object}
\usage{
initialize_omicselector_results(methods, parameters, de_results, start_time)
}
\arguments{
\item{methods}{Integer vector of methods}

\item{parameters}{List of validated parameters}

\item{de_results}{Data frame of DE results}

\item{start_time}{POSIXct start time}
}
\value{
List structure for results
}
\description{
Initialize OmicSelector Results Object
}
\keyword{internal}
