% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics-select-main.R
\name{initialize_results_object}
\alias{initialize_results_object}
\title{Initialize Results Object}
\usage{
initialize_results_object(stamp, config, methods, data_list)
}
\arguments{
\item{stamp}{Character string with run identifier}

\item{config}{Configuration list}

\item{methods}{Integer vector with methods}

\item{data_list}{List with loaded data}
}
\value{
OmicSelector results object
}
\description{
Initialize Results Object
}
\keyword{internal}
