% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{list_deprecated_functions}
\alias{list_deprecated_functions}
\title{List all deprecated functions}
\usage{
list_deprecated_functions()
}
\value{
A data.frame with deprecated function info
}
\description{
Returns a list of deprecated functions and their replacements
}
