% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-benchmarking.R
\name{load_benchmark_data}
\alias{load_benchmark_data}
\title{Load Benchmark Data}
\usage{
load_benchmark_data(wd)
}
\arguments{
\item{wd}{Working directory}
}
\value{
List with train, test, and optional validation data
}
\description{
Load Benchmark Data
}
\keyword{internal}
