% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-mirna-correction.R
\name{load_mirbase_aliases}
\alias{load_mirbase_aliases}
\title{Load miRBase Aliases Data}
\usage{
load_mirbase_aliases(
  species,
  use_cache,
  cache_dir,
  timeout_sec,
  verbose,
  fallback_file
)
}
\description{
Load miRBase Aliases Data
}
\keyword{internal}
