% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics-select-main.R
\name{load_omics_data}
\alias{load_omics_data}
\title{Load Omics Data Files}
\usage{
load_omics_data(wd)
}
\arguments{
\item{wd}{Working directory path}
}
\value{
List containing loaded data frames
}
\description{
Load Omics Data Files
}
\keyword{internal}
