% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianTuner.R
\name{make_autotuner_glmnet}
\alias{make_autotuner_glmnet}
\title{Create Bayesian-Optimized AutoTuner for glmnet}
\usage{
make_autotuner_glmnet(
  task,
  n_evals = NULL,
  inner_folds = 3,
  measure = "classif.auc"
)
}
\arguments{
\item{task}{An mlr3 Task}

\item{n_evals}{Number of Bayesian optimization iterations}

\item{inner_folds}{Number of inner CV folds}

\item{measure}{Performance measure}
}
\value{
An AutoTuner object
}
\description{
Creates an AutoTuner with glmnet (elastic net) and search space for alpha.
}
