% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianTuner.R
\name{make_autotuner_lightgbm}
\alias{make_autotuner_lightgbm}
\title{Create Bayesian-Optimized AutoTuner for LightGBM}
\usage{
make_autotuner_lightgbm(
  task,
  n_evals = NULL,
  inner_folds = 3,
  measure = "classif.auc",
  nthread = 4L
)
}
\arguments{
\item{task}{An mlr3 Task}

\item{n_evals}{Number of Bayesian optimization iterations}

\item{inner_folds}{Number of inner CV folds}

\item{measure}{Performance measure}

\item{nthread}{Number of threads}
}
\value{
An AutoTuner object
}
\description{
Creates an AutoTuner with LightGBM learner and omics-optimized search space.
}
