% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianTuner.R
\name{make_autotuner_ranger}
\alias{make_autotuner_ranger}
\title{Create Bayesian-Optimized AutoTuner for Random Forest}
\usage{
make_autotuner_ranger(
  task,
  n_evals = NULL,
  inner_folds = 3,
  measure = "classif.auc",
  num_trees = 500L
)
}
\arguments{
\item{task}{An mlr3 Task}

\item{n_evals}{Number of Bayesian optimization iterations}

\item{inner_folds}{Number of inner CV folds}

\item{measure}{Performance measure}

\item{num_trees}{Number of trees (fixed, usually not worth tuning)}
}
\value{
An AutoTuner object
}
\description{
Creates an AutoTuner with ranger and omics-optimized search space.
}
