% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterGOF.R
\name{make_gof_filter}
\alias{make_gof_filter}
\title{Create GOF Filter}
\usage{
make_gof_filter(type = c("ks", "hurdle", "zero_prop"), ...)
}
\arguments{
\item{type}{Filter type: "ks", "hurdle", or "zero_prop"}

\item{...}{Additional parameters passed to the filter}
}
\value{
A Filter object
}
\description{
Convenience function to create GOF filters.
}
\examples{
\dontrun{
filt <- make_gof_filter("ks")
filt <- make_gof_filter("hurdle", min_nonzero_per_class = 5)
}

}
