% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio-image-cnn.R
\name{make_ratio_image}
\alias{make_ratio_image}
\title{Create Pairwise Ratio Image from Expression Vector}
\usage{
make_ratio_image(x)
}
\arguments{
\item{x}{Numeric vector of feature values on log scale.}
}
\value{
A square numeric matrix with one row and one column per feature.
}
\description{
Alias for [encode_ratio_image()] retained for backwards compatibility.
}
\examples{
x <- c(5, 3, 1)
make_ratio_image(x)

}
\references{
Aitchison J. (1986). \emph{The Statistical Analysis of Compositional Data}.
Chapman and Hall.
}
