% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio-image-cnn.R
\name{make_ratio_images}
\alias{make_ratio_images}
\title{Batch-Create Ratio Images from an Expression Matrix}
\usage{
make_ratio_images(mat)
}
\arguments{
\item{mat}{Numeric matrix with samples in rows and features in columns.}
}
\value{
A 3D array with dimensions `samples x features x features`.
}
\description{
Applies [encode_ratio_image()] row-wise to a log-scale expression matrix.
}
\examples{
mat <- matrix(c(5, 3, 1, 4, 2, 0), nrow = 2, byrow = TRUE)
imgs <- make_ratio_images(mat)
dim(imgs)

}
\references{
Aitchison J. (1986). \emph{The Statistical Analysis of Compositional Data}.
Chapman and Hall.
}
