% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-selection-pipeline.R
\name{method_boruta}
\alias{method_boruta}
\title{Method 17: Boruta}
\usage{
method_boruta(data, config)
}
\arguments{
\item{data}{Training data}

\item{config}{Configuration}
}
\value{
Method result
}
\description{
Method 17: Boruta
}
\keyword{internal}
