% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-selection-pipeline.R
\name{method_rf_rfe}
\alias{method_rf_rfe}
\title{Method 11: Random Forest RFE}
\usage{
method_rf_rfe(data, config)
}
\arguments{
\item{data}{Training data}

\item{config}{Configuration}
}
\value{
Method result
}
\description{
Method 11: Random Forest RFE
}
\keyword{internal}
