% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-selection-pipeline.R
\name{method_varselrf}
\alias{method_varselrf}
\title{Method 20: Variable Selection Random Forest}
\usage{
method_varselrf(data, config)
}
\arguments{
\item{data}{Training data}

\item{config}{Configuration}
}
\value{
Method result
}
\description{
Method 20: Variable Selection Random Forest
}
\keyword{internal}
