% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkService.R
\name{omic_benchmark}
\alias{omic_benchmark}
\title{Create benchmark service from OmicPipeline}
\usage{
omic_benchmark(pipeline, outer_folds = 5, inner_folds = 3, ...)
}
\arguments{
\item{pipeline}{An OmicPipeline object}

\item{outer_folds}{Number of outer CV folds}

\item{inner_folds}{Number of inner CV folds}

\item{...}{Additional arguments passed to BenchmarkService$new()}
}
\value{
A BenchmarkService object
}
\description{
Convenience function to create a BenchmarkService
}
