% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-system.R
\name{omics_log}
\alias{omics_log}
\title{Internal logging function}
\usage{
omics_log(level, message, ...)
}
\arguments{
\item{level}{Log level}

\item{message}{Log message}

\item{...}{Additional parameters}
}
\description{
Internal logging function
}
\keyword{internal}
