% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_combat.R
\name{OmicSelector_combat}
\alias{OmicSelector_combat}
\title{OmicSelector_combat}
\usage{
OmicSelector_combat(
  danex,
  metadane = metadane,
  model = c("~ Batch", "~ Batch + Class")
)
}
\arguments{
\item{danex}{Matrix with miRNA normalized expression values (e.g. TPM, deltaCt) with miRNAs in columns and cases in rows.}

\item{metadane}{Metadata with `Class` and `Batch` variable.}

\item{model}{Model of correction (application of covariates).}
}
\value{
Batch-corrected dataset.
}
\description{
Use combat to fight batch effect.
}
