% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_correct_miRNA_names.R,
%   R/modern-mirna-correction.R
\name{OmicSelector_correct_miRNA_names}
\alias{OmicSelector_correct_miRNA_names}
\title{OmicSelector_correct_miRNA_names}
\usage{
OmicSelector_correct_miRNA_names(temp, species = "hsa", correct_dots = TRUE)

OmicSelector_correct_miRNA_names(temp, species = "hsa", correct_dots = TRUE)
}
\arguments{
\item{temp}{Dataset with miRNA names in columns.}

\item{correct_dots}{Boolean variable to correct the names after correction of dots to hyphens. This tries to compensate the effect of make.names() function.}
}
\value{
Corrected dataset.
}
\description{
Sometinmes, when using the dataset mapped to previous versions of miRbase we may get false mismatches due to changes in terminology.
This function uses latest version of miRbase to correct all old miRNA names to new one.
}
