% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_counts_to_log10tpm.R
\name{OmicSelector_counts_to_log10tpm}
\alias{OmicSelector_counts_to_log10tpm}
\title{OmicSelector_counts_to_log10tpm}
\usage{
OmicSelector_counts_to_log10tpm(
  danex,
  metadane = metadane,
  ids = metadane$ID,
  filtr = T,
  filtr_minimalcounts = 10,
  filtr_howmany = 1/2
)
}
\arguments{
\item{danex}{Matrix with miRNA counts with miRNAs in columns and cases in rows.}

\item{metadane}{Metadata with `Class` variable.}

\item{ids}{Unique identifier of samples.}

\item{filtr}{If expression filter should be used.}

\item{filtr_minimalcounts}{How many counts?}

\item{filtr_howmany}{In how many samples? (Please provide a percentage or proportion, e.g. 1/2 or 1/3).}
}
\value{
Normalized counts in `ttpm` format. Please note, that the function also saves `TPM_DGEList_filtered.rds` to working directory, which is a DGEList object that can be used in packages like edgeR.
}
\description{
Counts to log-transformed TPM-normalized counts.
The funcction support additional filter, i.e. it can leave miRNAs that appear in minimum of `filtr_minimalcounts` counts in `filtr_howmany` samples.
Usage of the filter like that can assure that the miRNAs selected will be detectable in qPCR.
}
