% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_create_formula.R
\name{OmicSelector_create_formula}
\alias{OmicSelector_create_formula}
\title{OmicSelector_create_formula}
\usage{
OmicSelector_create_formula(selected_features)
}
\arguments{
\item{selected_features}{Selected miRNAs as characted vector.}
}
\value{
Formula "Class ~ ..." to be used in another functions.
}
\description{
Helper function to create formula based on selected miRNAs.
}
