% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_diverge_color.R
\name{OmicSelector_diverge_color}
\alias{OmicSelector_diverge_color}
\title{OmicSelector_diverge_color}
\usage{
OmicSelector_diverge_color(data, centeredOn = 0)
}
\description{
Helper for plotting.
}
