% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_docker.update_progress.R
\name{OmicSelector_docker.update_progress}
\alias{OmicSelector_docker.update_progress}
\title{OmicSelector_docker.update_progress}
\usage{
OmicSelector_docker.update_progress(verbose = F)
}
\description{
Helper for docker. Updates progress of analysis by 1 step.
}
