% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/OmicSelector_download_tissue_miRNA_data_from_TCGA.R
\name{OmicSelector_download_tissue_miRNA_data_from_TCGA}
\alias{OmicSelector_download_tissue_miRNA_data_from_TCGA}
\title{OmicSelector_download_tissue_miRNA_data_from_TCGA}
\usage{
OmicSelector_download_tissue_miRNA_data_from_TCGA(data_folder = getwd())
}
\arguments{
\item{data_folder}{Path. Where to dowload the data?}
}
\description{
Function allowing to dowload all miRNA-seq data from TCGA.
}
