% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_get_benchmark.R
\name{OmicSelector_get_benchmark}
\alias{OmicSelector_get_benchmark}
\title{OmicSelector_get_benchmark}
\usage{
OmicSelector_get_benchmark(benchmark_csv = "benchmark1578929876.21765.csv")
}
\arguments{
\item{benchmark_csv}{Path to csv file.}
}
\value{
Benchmark data frame.
}
\description{
Load benchmark results from csv file.
}
