% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_get_benchmark_methods.R
\name{OmicSelector_get_benchmark_methods}
\alias{OmicSelector_get_benchmark_methods}
\title{OmicSelector_get_benchmark_methods}
\usage{
OmicSelector_get_benchmark_methods(benchmark_csv = "benchmark1578929876.21765.csv")
}
\arguments{
\item{benchmark_csv}{Path to benchmark csv file.}
}
\value{
Vector of feature selection methods checked.
}
\description{
Get methods checked in benchmark.
}
