% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_heatmap.R
\name{OmicSelector_heatmap}
\alias{OmicSelector_heatmap}
\title{OmicSelector_heatmap}
\usage{
OmicSelector_heatmap(
  x = trainx[, 1:10],
  rlab = data.frame(Batch = dane$Batch, Class = dane$Class),
  zscore = F,
  margins = c(10, 10),
  expression_name = "log10(TPM)",
  trim_min = NULL,
  trim_max = NULL,
  centered_on = NULL,
  legend_pos = "topright",
  legend_cex = 0.8,
  ...
)
}
\arguments{
\item{x}{Matrix of log-transformed TPM-normalized counts with miRNAs in columns and cases in rows.}

\item{rlab}{Data frame of factors to be marked on heatmap (like batch or class). Maximum of 2 levels for every variable is supported.}

\item{zscore}{Whether to z-score values before clustering and plotting.}

\item{expression_name}{What should be written on the plot?}

\item{trim_min}{Trim lower than.. Useful for setting appropriate scale}

\item{trim_max}{Trim greater than.. Useful for setting appropriate scale}

\item{centered_on}{On which value should the scale be centered? If null - median will be used.}

\item{legend_pos}{Where should the legend should be? Default: topright}

\item{legend_cex}{How large should the legend should be? Default: 0.8}
}
\value{
Heatmap.
}
\description{
Draw a heatmap of selected miRNAs.
}
