% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_iteratedRFE.R
\name{OmicSelector_iteratedRFE}
\alias{OmicSelector_iteratedRFE}
\title{OmicSelector_iteratedRFE}
\usage{
OmicSelector_iteratedRFE(
  trainSet,
  testSet = NULL,
  initFeatures = colnames(trainSet),
  classLab,
  checkNFeatures = 25,
  votingIterations = 1000,
  useCV = F,
  nfolds = 10,
  initRandomState = 42
)
}
\description{
Helper for Marcin Kaszkowiak propritary method.
Very longs.. needs optimizing.
}
