% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_load_datamix.R
\name{OmicSelector_load_datamix}
\alias{OmicSelector_load_datamix}
\title{OmicSelector_load_datamix}
\usage{
OmicSelector_load_datamix(
  wd = getwd(),
  smote_over = 10000,
  use_smote_not_rose = T,
  replace_smote = F,
  selected_miRNAs = NULL
)
}
\arguments{
\item{wd}{Working directory with files for the loading.}

\item{smote_over}{Oversampling of minority class in SMOTE function (deterimes the number of cases in final dataset). See `perc.over` in `DMwR::SMOTE()`` function.}

\item{use_smote_not_rose}{Set TRUE for SMOTE instead of ROSE.}

\item{replace_smote}{For some analyses we may want to replace imbalanced train dataset with balanced dataset. This saved coding time in some functions.}

\item{selected_miRNAs}{If null - take all features staring with "hsa", if set - vector of feature names to be selected.}
}
\value{
The list of objects in the following order: train, test, valid, train_smoted, trainx, trainx_smoted, merged. (trainx contains only the miRNA data without metadata)
}
\description{
This function loads the data created in preparation phase.
It requires the output constructed by `OmicSelector_prepare_split` function to be placed in working directory (`wd`), thus files `mixed_train.csv`, `mixed_test.csv` and `mixed_valid.csv` have to exist in the directory.
For imbalanced data, the fuction can perform balancing using:
1. ROSE: https://journal.r-project.org/archive/2014/RJ-2014-008/RJ-2014-008.pdf - by default we generate 10 * number of cases in orginal dataset.
2. SMOTE (default): https://arxiv.org/abs/1106.1813 - by defult we use `perc.under=100` and `k=10`.
}
