% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_load_extension.R
\name{OmicSelector_load_extension}
\alias{OmicSelector_load_extension}
\title{OmicSelector_load_extension}
\usage{
OmicSelector_load_extension(name = "deeplearning")
}
\arguments{
\item{name}{Name of extension to be loaded.}
}
\description{
Function for loading extension.
}
