% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_log.R
\name{OmicSelector_log}
\alias{OmicSelector_log}
\title{OmicSelector_log}
\usage{
OmicSelector_log(message_to_log = "Working...", logfile = "temp.log")
}
\arguments{
\item{logfile}{Path to logfile.}

\item{message}{Message to log.}
}
\description{
Logging to sink.
}
