% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_differential_expression_ttest.R
\name{OmicSelector_differential_expression_ttest}
\alias{OmicSelector_differential_expression_ttest}
\title{OmicSelector_differential_expression_ttest}
\usage{
OmicSelector_differential_expression_ttest(ttpm_features, classes, mode = "auto")
}
\arguments{
\item{ttpm_features}{matrix of log-transformed TPM-normalized miRNAs counts with miRNAs placed in `ttpm` design (i.e. columns and cases placed as rows)}

\item{classes}{vector describing label for each case. It should contain only "Cancer" and "Control" labeles!!!!}

\item{mode}{use 'logtpm' for log(TPM) data or 'deltact' for qPCR deltaCt values. This parameters sets how the fold-change is calculated. Setting it to "auto" will try to read settings from var_type.txt (used in docker).}
}
\value{
Data frame with results.
}
\description{
The variable performes standard differential expression analysis using unpaired t-test with BH and Bonferonni correction.
It requires `ttpm_features` object, which is a matrix of log-transformed TPM-normalized miRNAs counts with miRNAs placed in `ttpm` design (i.e. columns and cases placed as rows).
Classess should be passed as `classes` and this should be a vector of length equal to number of rows in `ttpm_polfiltrze` and contain only "Cancer" or "Control" labels!!
The function returns the miRNAs sorted by BH-corrected p-value.
}
