% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_signiture_overlap.R
\name{OmicSelector_signiture_overlap}
\alias{OmicSelector_signiture_overlap}
\title{OmicSelector_signiture_overlap}
\usage{
OmicSelector_signiture_overlap(
  which_formulas = c("sig", "cfs"),
  benchmark_csv = "benchmark1578929876.21765.csv"
)
}
\arguments{
\item{which_formulas}{Which formulas to check?}

\item{benchmark_csv}{Which benchmark to use?}
}
\value{
Object of `venn()` function which can be used for plotting venn diagram and check the overlap.
}
\description{
A function to generate venn diagram and check the overlap between formulas.
}
