% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_myclust.R
\name{OmicSelector_myclust}
\alias{OmicSelector_myclust}
\title{OmicSelector_myclust}
\usage{
OmicSelector_myclust(c)
}
\description{
Helper in heatmap creation. Which method of cultering should be used?
}
