% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_mydist.R
\name{OmicSelector_mydist}
\alias{OmicSelector_mydist}
\title{OmicSelector_mydist}
\usage{
OmicSelector_mydist(c)
}
\description{
Helper in heatmap creation. Which distance should be used?
}
