% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_PCA_3D.R
\name{OmicSelector_PCA_3D}
\alias{OmicSelector_PCA_3D}
\title{OmicSelector_PCA_3D}
\usage{
OmicSelector_PCA_3D(ttpm_features, meta)
}
\arguments{
\item{ttpm_features}{Normalizaed counts in `ttpm` format.}

\item{meta}{Factor of cases labels that should be visualized on biplot.}
}
\value{
Plotly 3D object.
}
\description{
Conduct PCA and create 3D scatterplot form first 3 components.
}
