% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_setup.R
\name{OmicSelector_setup}
\alias{OmicSelector_setup}
\title{OmicSelector_setup}
\usage{
OmicSelector_setup(keras = TRUE, msg = TRUE)
}
\arguments{
\item{keras}{Try to install keras using install_keras()?}

\item{msg}{Display OK message at the end?}
}
\description{
Run this function to be sure that everything is installed properly for OmicSelector.
}
