% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_table.R
\name{OmicSelector_table}
\alias{OmicSelector_table}
\title{OmicSelector_table}
\usage{
OmicSelector_table(table, height = "400px", ...)
}
\arguments{
\item{table}{Table to draw.}

\item{hight}{High (default: 400px)}
}
\description{
A wrapper to draw pretty tables in rmarkdown document
}
