% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_vulcano_plot.R
\name{OmicSelector_vulcano_plot}
\alias{OmicSelector_vulcano_plot}
\title{OmicSelector_vulcano_plot}
\usage{
OmicSelector_vulcano_plot(
  selected_miRNAs,
  DE,
  only_label = NULL,
  take_adjusted_p = FALSE,
  log2fc_col = "log2FC",
  p_col = "p-value",
  p_adj_col = "p-value BH",
  title = NULL
)
}
\arguments{
\item{selected_miRNAs}{Vector of selected miRNAs.}

\item{DE}{Differential expression data.frame containing columns for log2FC and p-values.}

\item{only_label}{If NULL - all selected miRNAs will be labeled. If you want to label only selected miRNAs provide a character vector with their names.}

\item{take_adjusted_p}{Use adjusted p-values (BH) if TRUE.}

\item{log2fc_col}{Column name for log2 fold-change (default: "log2FC").}

\item{p_col}{Column name for raw p-values (default: "p-value").}

\item{p_adj_col}{Column name for adjusted p-values (default: "p-value BH").}

\item{title}{Optional plot title.}
}
\value{
A ggplot object.
}
\description{
Draw a volcano plot of selected features from a differential expression table.
}
