% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_xgboost.R
\name{OmicSelector_xgboost}
\alias{OmicSelector_xgboost}
\title{OmicSelector_xgboost}
\usage{
OmicSelector_xgboost(
  features = "all",
  train = OmicSelector_load_datamix(use_smote_not_rose = T)[[1]],
  test = OmicSelector_load_datamix(use_smote_not_rose = T)[[2]],
  valid = OmicSelector_load_datamix(use_smote_not_rose = T)[[2]],
  eta = c(0, 1),
  gamma = c(0, 100),
  max_depth = c(2L, 10L),
  min_child_weight = c(1, 25),
  subsample = c(0.25, 1),
  nfold = c(3L, 10L),
  initPoints = 8,
  iters.n = 10
)
}
\arguments{
\item{features}{Vector of features to be used. If "all", all features starting with 'hsa' will be used.}

\item{train}{Training dataset with column Class ('Case' vs. 'Control') and features starting with 'hsa'.}

\item{test}{Testing dataset with column Class ('Case' vs. 'Control') and features starting with 'hsa'.}

\item{valid}{Testing dataset with column Class ('Case' vs. 'Control') and features starting with 'hsa'.}

\item{eta}{Bonderies of 'eta' parameter in XGBoost training, must be a vector of 2.}

\item{gamma}{Bonderies of 'gamma' parameter in XGBoost training, must be a vector of 2.}

\item{max_depth}{Bonderies of 'max_depth' parameter in XGBoost training, must be a vector of 2.}

\item{min_child_weight}{Bonderies of 'min_child_weight' parameter in XGBoost training, must be a vector of 2.}

\item{subsample}{Bonderies of 'subsample' parameter in XGBoost training, must be a vector of 2.}

\item{nfold}{Bonderies of 'nfold' parameter in XGBoost training, must be a vector of 2.}
}
\value{
Xgboost model
}
\description{
Train xgboost model with Bayesian optimalization.
Code based on http://www.mysmu.edu/faculty/jwwang/post/hyperparameters-tuning-for-xgboost-using-bayesian-optimization/
}
