% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlesample-deprecated-aliases.R
\name{paper3-deprecated}
\alias{paper3-deprecated}
\alias{paper3_assert_method_bank_exports}
\alias{paper3_assert_row_equivariant}
\alias{paper3_bh_fdr_correct_blocked}
\alias{paper3_bh_fdr_correct_matched_null}
\alias{paper3_hanley_mcneil_auc_ci}
\alias{paper3_holm_correct_familywise}
\alias{paper3_is_row_equivariant}
\alias{paper3_make_loco_splits}
\alias{paper3_make_locto_splits}
\alias{paper3_matched_null_benchmark}
\alias{paper3_matched_null_benchmark_cv}
\alias{paper3_method_bank}
\alias{paper3_method_roster}
\alias{paper3_paired_auc_diff_se}
\alias{paper3_register_score_adapter}
\alias{paper3_score_call}
\alias{paper3_technology_lift_delong}
\title{Deprecated `paper3_*` aliases}
\usage{
paper3_assert_method_bank_exports(...)

paper3_assert_row_equivariant(...)

paper3_bh_fdr_correct_blocked(...)

paper3_bh_fdr_correct_matched_null(...)

paper3_hanley_mcneil_auc_ci(...)

paper3_holm_correct_familywise(...)

paper3_is_row_equivariant(...)

paper3_make_loco_splits(...)

paper3_make_locto_splits(...)

paper3_matched_null_benchmark(...)

paper3_matched_null_benchmark_cv(...)

paper3_method_bank(...)

paper3_method_roster(...)

paper3_paired_auc_diff_se(...)

paper3_register_score_adapter(...)

paper3_score_call(...)

paper3_technology_lift_delong(...)
}
\arguments{
\item{...}{Arguments passed on to the corresponding `singlesample_*` function.}
}
\value{
The value of the corresponding `singlesample_*` function.
}
\description{
Back-compatibility aliases for the renamed `singlesample_*` functions. Each
forwards verbatim to its `singlesample_*` counterpart and is retained so that
code written against the former `paper3_*` names keeps working. Prefer the
`singlesample_*` names; the `paper3_*` aliases may be removed in a future
major release.
}
\keyword{internal}
