% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-differential-expression.R
\name{perform_de_analysis}
\alias{perform_de_analysis}
\title{Perform Differential Expression Analysis}
\usage{
perform_de_analysis(
  processed_data,
  data_type,
  p_adjust_methods,
  effect_size_methods,
  parallel,
  verbose
)
}
\description{
Perform Differential Expression Analysis
}
\keyword{internal}
