% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-differential-expression.R
\name{perform_tests_parallel}
\alias{perform_tests_parallel}
\title{Perform Statistical Tests in Parallel}
\usage{
perform_tests_parallel(
  expression_data,
  case_idx,
  control_idx,
  data_type,
  verbose
)
}
\description{
Perform Statistical Tests in Parallel
}
\keyword{internal}
