% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-differential-expression.R
\name{perform_tests_sequential}
\alias{perform_tests_sequential}
\title{Perform Statistical Tests Sequentially}
\usage{
perform_tests_sequential(
  expression_data,
  case_idx,
  control_idx,
  data_type,
  verbose
)
}
\description{
Perform Statistical Tests Sequentially
}
\keyword{internal}
