% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{plot.OmicSelector}
\alias{plot.OmicSelector}
\title{Plot Method for OmicSelector Objects}
\usage{
\method{plot}{OmicSelector}(x, type = "overview", ...)
}
\arguments{
\item{x}{OmicSelector object}

\item{type}{Character string specifying plot type}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot Method for OmicSelector Objects
}
