% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{plot_feature_counts}
\alias{plot_feature_counts}
\title{Plot Feature Count Distribution}
\usage{
plot_feature_counts(x, ...)
}
\arguments{
\item{x}{OmicSelector object}

\item{...}{Additional arguments}
}
\description{
Plot Feature Count Distribution
}
\keyword{internal}
