% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{plot_method_success}
\alias{plot_method_success}
\title{Plot Method Success Rate}
\usage{
plot_method_success(x, ...)
}
\arguments{
\item{x}{OmicSelector object}

\item{...}{Additional arguments}
}
\description{
Plot Method Success Rate
}
\keyword{internal}
