% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{plot_overview}
\alias{plot_overview}
\title{Plot Overview of Results}
\usage{
plot_overview(x, ...)
}
\arguments{
\item{x}{OmicSelector object}

\item{...}{Additional arguments}
}
\description{
Plot Overview of Results
}
\keyword{internal}
