% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemolysis-correction.R
\name{predict_weighted_clr_mlp}
\alias{predict_weighted_clr_mlp}
\title{Predict with a Weighted CLR + MLP Classifier}
\usage{
predict_weighted_clr_mlp(fit, X_new)
}
\arguments{
\item{fit}{A fitted `"omicselector_weighted_clr_mlp"` object.}

\item{X_new}{Numeric matrix of new samples.}
}
\value{
Numeric vector of positive-class probabilities.
}
\description{
Generates positive-class probabilities for new samples from a fit created by
[train_weighted_clr_mlp()].
}
