% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-differential-expression.R
\name{prepare_de_data}
\alias{prepare_de_data}
\title{Prepare Data for Differential Expression}
\usage{
prepare_de_data(
  expression_data,
  class_labels,
  case_label,
  control_label,
  data_type,
  verbose
)
}
\description{
Prepare Data for Differential Expression
}
\keyword{internal}
