% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{print.CalibrationResult}
\alias{print.CalibrationResult}
\title{Print method for CalibrationResult}
\usage{
\method{print}{CalibrationResult}(x, ...)
}
\arguments{
\item{x}{CalibrationResult object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for CalibrationResult
}
