% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{print.LeakageValidation}
\alias{print.LeakageValidation}
\title{Print method for LeakageValidation}
\usage{
\method{print}{LeakageValidation}(x, ...)
}
\arguments{
\item{x}{A LeakageValidation object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for LeakageValidation
}
