% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-normalization.R
\name{print.mirna_normalized}
\alias{print.mirna_normalized}
\title{Print method for mirna_normalized objects}
\usage{
\method{print}{mirna_normalized}(x, ...)
}
\description{
Print method for mirna_normalized objects
}
