% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkService.R
\name{print.NestedCVResult}
\alias{print.NestedCVResult}
\title{Print method for NestedCVResult}
\usage{
\method{print}{NestedCVResult}(x, ...)
}
\arguments{
\item{x}{A NestedCVResult object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for NestedCVResult
}
