% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{print.NogueiraStability}
\alias{print.NogueiraStability}
\title{Print method for NogueiraStability}
\usage{
\method{print}{NogueiraStability}(x, ...)
}
\arguments{
\item{x}{A NogueiraStability object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for NogueiraStability
}
