% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicPipeline.R
\name{print.OmicBenchmarkResult}
\alias{print.OmicBenchmarkResult}
\title{Print method for OmicBenchmarkResult}
\usage{
\method{print}{OmicBenchmarkResult}(x, ...)
}
\arguments{
\item{x}{An OmicBenchmarkResult object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for OmicBenchmarkResult
}
