% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{print.OmicSelector}
\alias{print.OmicSelector}
\title{Print Method for OmicSelector Objects}
\usage{
\method{print}{OmicSelector}(x, ...)
}
\arguments{
\item{x}{OmicSelector object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for OmicSelector Objects
}
