% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_OmicSelector_improved.R
\name{print.omicselector_legacy}
\alias{print.omicselector_legacy}
\title{Print Method for Legacy OmicSelector Results}
\usage{
\method{print}{omicselector_legacy}(x, ...)
}
\arguments{
\item{x}{OmicSelector legacy results object}

\item{...}{Additional arguments}
}
\description{
Print Method for Legacy OmicSelector Results
}
