% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-optimized.R
\name{print.omicselector_results}
\alias{print.omicselector_results}
\title{Print Method for OmicSelector Results}
\usage{
\method{print}{omicselector_results}(x, ...)
}
\arguments{
\item{x}{OmicSelector results object}

\item{...}{Additional arguments}
}
\description{
Print Method for OmicSelector Results
}
