% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-omics.R
\name{print.OmicsInput}
\alias{print.OmicsInput}
\title{Print method for OmicsInput}
\usage{
\method{print}{OmicsInput}(x, ...)
}
\arguments{
\item{x}{An OmicsInput object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for OmicsInput
}
