% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print.ReportData}
\alias{print.ReportData}
\title{Print method for ReportData}
\usage{
\method{print}{ReportData}(x, ...)
}
\arguments{
\item{x}{A ReportData object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for ReportData
}
