% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlesample-deploy.R
\name{print.singlesample_deployable}
\alias{print.singlesample_deployable}
\title{Print a single-sample deployment object}
\usage{
\method{print}{singlesample_deployable}(x, ...)
}
\arguments{
\item{x}{A \code{singlesample_deployable} object.}

\item{...}{Unused.}
}
\value{
Invisibly, \code{x}.
}
\description{
Print a single-sample deployment object
}
