% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlesample-selector.R
\name{print.singlesample_selector}
\alias{print.singlesample_selector}
\title{Print a single-sample selector}
\usage{
\method{print}{singlesample_selector}(x, ...)
}
\arguments{
\item{x}{A `singlesample_selector` or
`singlesample_selector_ineligible` object.}

\item{...}{Unused.}
}
\value{
Invisibly, `x`.
}
\description{
Print a single-sample selector
}
