% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda-feature-selection.R
\name{register_coda_feature_selection_filters}
\alias{register_coda_feature_selection_filters}
\title{Register CoDA Feature Selection Filters}
\usage{
register_coda_feature_selection_filters()
}
\description{
Registers the CoDA-aware `mlr3filters` feature filters provided by this
package. The function is idempotent and is also called on package load.
}
