% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterGOF.R
\name{register_gof_filters}
\alias{register_gof_filters}
\title{Register GOF Filters in mlr3}
\usage{
register_gof_filters()
}
\description{
Registers the GOF filters in the mlr3filters dictionary.
Call this function to make filters available via `flt("gof_ks")`, etc.
}
