% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{reliability_diagram_data}
\alias{reliability_diagram_data}
\title{Create Reliability Diagram Data}
\usage{
reliability_diagram_data(probs, labels, n_bins = 10)
}
\arguments{
\item{probs}{Predicted probabilities}

\item{labels}{True labels}

\item{n_bins}{Number of bins}
}
\value{
A data frame suitable for ggplot2
}
\description{
Generates data for plotting a reliability (calibration) diagram.
}
\examples{
\dontrun{
library(ggplot2)

diagram_data <- reliability_diagram_data(probs, labels)

ggplot(diagram_data, aes(x = mean_predicted, y = fraction_positive)) +
  geom_point(aes(size = n_samples)) +
  geom_abline(slope = 1, intercept = 0, linetype = "dashed") +
  xlim(0, 1) + ylim(0, 1) +
  labs(x = "Mean Predicted Probability", y = "Fraction of Positives")
}

}
