% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern-benchmarking.R
\name{save_benchmark_results}
\alias{save_benchmark_results}
\title{Save Benchmark Results}
\usage{
save_benchmark_results(benchmark_results, wd, output_file)
}
\arguments{
\item{benchmark_results}{Benchmark results object}

\item{wd}{Working directory}

\item{output_file}{Output file name}
}
\description{
Save Benchmark Results
}
\keyword{internal}
