% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics-select-main.R
\name{save_results}
\alias{save_results}
\title{Save Results to Files}
\usage{
save_results(results, wd)
}
\arguments{
\item{results}{OmicSelector results object}

\item{wd}{Working directory}
}
\description{
Save Results to Files
}
\keyword{internal}
