% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{summary.OmicSelector}
\alias{summary.OmicSelector}
\title{Summary Method for OmicSelector Objects}
\usage{
\method{summary}{OmicSelector}(object, ...)
}
\arguments{
\item{object}{OmicSelector object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for OmicSelector Objects
}
