% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_OmicSelector_improved.R
\name{summary.omicselector_legacy}
\alias{summary.omicselector_legacy}
\title{Summary Method for Legacy OmicSelector Results}
\usage{
\method{summary}{omicselector_legacy}(object, ...)
}
\arguments{
\item{object}{OmicSelector legacy results object}

\item{...}{Additional arguments}
}
\description{
Summary Method for Legacy OmicSelector Results
}
