% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-optimized.R
\name{validate_omicselector_inputs}
\alias{validate_omicselector_inputs}
\title{Validate OmicSelector Input Parameters}
\usage{
validate_omicselector_inputs(
  data_path,
  methods,
  class_column,
  case_label,
  control_label,
  feature_prefix,
  max_features,
  max_iterations,
  p_threshold,
  fc_threshold,
  parallel_cores,
  timeout_minutes,
  output_dir
)
}
\arguments{
\item{...}{All input parameters from main function}
}
\value{
List of validated parameters
}
\description{
Validate OmicSelector Input Parameters
}
\keyword{internal}
