% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utils.R
\name{validate_working_directory}
\alias{validate_working_directory}
\title{Validate Working Directory}
\usage{
validate_working_directory(
  wd,
  required_files = c("mixed_train.csv", "mixed_test.csv")
)
}
\arguments{
\item{wd}{Character string with working directory path}

\item{required_files}{Character vector with required file names}
}
\value{
Logical, TRUE if valid
}
\description{
Validate Working Directory
}
\keyword{internal}
