% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{with_parallel}
\alias{with_parallel}
\title{With Parallel Scope}
\usage{
with_parallel(expr, workers = NULL, plan = "multisession")
}
\arguments{
\item{expr}{Expression to evaluate}

\item{workers}{Number of workers}

\item{plan}{Parallelization plan}
}
\value{
Result of expression
}
\description{
Execute code within a parallelization scope, automatically cleaning up.
}
