% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicselector-helpers.R
\name{withTimeout}
\alias{withTimeout}
\title{Safe Timeout Function}
\usage{
withTimeout(expr, timeout)
}
\arguments{
\item{expr}{Expression to execute}

\item{timeout}{Timeout in seconds}
}
\value{
Result of expression or timeout error
}
\description{
Safe Timeout Function
}
\keyword{internal}
