% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoXAI.R
\name{xai_pdp}
\alias{xai_pdp}
\title{Compute Partial Dependence Plots}
\usage{
xai_pdp(explainer, features, grid_n = 20, grid_type = "quantile")
}
\arguments{
\item{explainer}{A DALEX explainer}

\item{features}{Features to compute PDPs for}

\item{grid_n}{Number of grid points (default: 20)}

\item{grid_type}{"quantile" or "uniform"}
}
\value{
A model_profile object
}
\description{
Computes PDPs for specified features using DALEX.
}
