% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoXAI.R
\name{xai_shap}
\alias{xai_shap}
\title{Compute SHAP Values for Observations}
\usage{
xai_shap(explainer, new_observations, B = 25)
}
\arguments{
\item{explainer}{A DALEX explainer}

\item{new_observations}{Data frame of observations to explain}

\item{B}{Number of random orderings (default: 25)}
}
\value{
A list of predict_parts objects (one per observation)
}
\description{
Computes SHAP values for individual predictions.
}
