% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterGOF.R
\name{FilterZeroProp}
\alias{FilterZeroProp}
\title{Zero-Proportion Filter}
\description{
Simple filter based on the difference in zero-proportion between classes.
Useful for quickly identifying on/off switches in gene expression.
}
\details{
Computes: |P(X=0|Class1) - P(X=0|Class2)|
Higher values indicate features that are expressed in one class but not the other.
}
\section{Super class}{
\code{mlr3filters::Filter} -> \code{FilterZeroProp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilterZeroProp-new}{\code{FilterZeroProp$new()}}
\item \href{#method-FilterZeroProp-clone}{\code{FilterZeroProp$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="calculate"><a href='../../mlr3filters/html/Filter.html#method-Filter-calculate'><code>mlr3filters::Filter$calculate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="format"><a href='../../mlr3filters/html/Filter.html#method-Filter-format'><code>mlr3filters::Filter$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="help"><a href='../../mlr3filters/html/Filter.html#method-Filter-help'><code>mlr3filters::Filter$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="print"><a href='../../mlr3filters/html/Filter.html#method-Filter-print'><code>mlr3filters::Filter$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterZeroProp-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilterZeroProp-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this Filter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterZeroProp$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterZeroProp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilterZeroProp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterZeroProp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
