% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_correlation_plot.R
\name{OmicSelector_correlation_plot}
\alias{OmicSelector_correlation_plot}
\title{OmicSelector_correlation_plot}
\usage{
OmicSelector_correlation_plot(
  var1,
  var2,
  labvar1,
  labvar2,
  title = NULL,
  yx = TRUE,
  metoda = "pearson",
  gdzie_legenda = "topleft"
)
}
\arguments{
\item{var1}{First numeric variable.}

\item{var2}{Second numeric variable.}

\item{labvar1}{Label to be written on plot for fist variable.}

\item{labvar2}{Label to be written on plot for second variable.}

\item{title}{Optional plot title.}

\item{yx}{Logical varable. TRUE if you want f(x)=x axis for reference (usefull for calibration plots).}

\item{metoda}{Correlation method: "pearson" or "spearman".}

\item{gdzie_legenda}{Where the legend should be placed?}
}
\description{
Draw a correlation scatterplot with optional y = x reference and summary stats.
}
