% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_merge_formulas.R
\name{OmicSelector_merge_formulas}
\alias{OmicSelector_merge_formulas}
\title{OmicSelector_merge_formulas}
\usage{
OmicSelector_merge_formulas(wd = getwd(), max_miRNAs = 11, add = list())
}
\arguments{
\item{wd}{Working directory with formulas*.RDS files.}

\item{max_miRNAs}{Maximum number of miRNAs to be selected in formulas.}

\item{add}{List of additional sets that should be added to the formulas.}
}
\value{
Final formulas object, also save as `featureselection_formulas_final.RDS` in working directory.
}
\description{
Merge and filter formulas.
This function can be used to merge the formulas*.RDS created by different runs of `OmicSelector_OmicSelector()` and filter them to keep the maximum number of miRNAs.
This may be useful in planning qPCR validation.
The result of this function is `featureselection_formulas_final.RDS` file, which can be futher supplied to `OmicSelector_benchmark()`.
}
