% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_PCA.R
\name{OmicSelector_PCA}
\alias{OmicSelector_PCA}
\title{OmicSelector_PCA}
\usage{
OmicSelector_PCA(ttpm_features, meta)
}
\arguments{
\item{ttpm_features}{Normalizaed counts in `ttpm` format.}

\item{meta}{Factor of cases labels that should be visualized on biplot.}
}
\value{
Biplot.
}
\description{
Conduct PCA and create biplot.
}
