% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_profileplot.R
\name{OmicSelector_profileplot}
\alias{OmicSelector_profileplot}
\title{OmicSelector_profileplot}
\usage{
OmicSelector_profileplot(
  form,
  Method.id,
  standardize = TRUE,
  interval = 10,
  by.pattern = TRUE,
  original.names = TRUE
)
}
\description{
Modified function of profileR::profileplot to correct "Method".
}
