% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_propensity_score_matching.R
\name{OmicSelector_propensity_score_matching}
\alias{OmicSelector_propensity_score_matching}
\title{OmicSelector_propensity_score_matching}
\usage{
OmicSelector_propensity_score_matching(
  dataset,
  match_by = c("age_at_diagnosis", "gender.x"),
  method = "nearest",
  distance = "logit"
)
}
\arguments{
\item{dataset}{Original dataset.}

\item{match_by}{Vector describing by which variables should the dataset be matched.}

\item{method}{Passed to `matchit()`.}

\item{distance}{Passed to `matchit()`.}
}
\description{
Propensity score matching.
}
