% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSelector_vulcano_plot.R
\name{OmicSelector_vulcano_plot}
\alias{OmicSelector_vulcano_plot}
\title{OmicSelector_vulcano_plot}
\usage{
OmicSelector_vulcano_plot(
  selected_miRNAs,
  DE = OmicSelector_differential_expression_ttest(),
  only_label = NULL,
  take_adjusted_p = FALSE
)
}
\arguments{
\item{selected_miRNAs}{Vector of selected miRNAs.}

\item{DE}{Differential expression to get the values from. It can be generated by `OmicSelector_differential_expression_ttest()` or `OmicSelector_best_signature_de()`}

\item{only_label}{If NULL - all selected miRNAs will be labeled. If you want to label only selected miRNAs provide a character vector with their names.}

\item{take_adjusted_p}{If take adjusted p-value using BH method. Default: false}
}
\value{
Vulcano plot.
}
\description{
Draw vulcano plot of selected miRNAs.
}
