% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretability.R
\name{print.CorrelationCheck}
\alias{print.CorrelationCheck}
\title{Print Correlation Check}
\usage{
\method{print}{CorrelationCheck}(x, ...)
}
\arguments{
\item{x}{CorrelationCheck object to print.}

\item{...}{Unused; present for method compatibility.}
}
\description{
Print Correlation Check
}
