% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretability.R
\name{print.FeatureImportance}
\alias{print.FeatureImportance}
\title{Print Feature Importance}
\usage{
\method{print}{FeatureImportance}(x, n = 10, ...)
}
\arguments{
\item{x}{A `FeatureImportance` object.}

\item{n}{Number of top features to print.}

\item{...}{Additional arguments passed through to base methods (ignored).}
}
\description{
Print Feature Importance
}
